### 2021-07-21 #3517 @whedon
[PRE REVIEW]: WODEN: A CUDA-enabled package to simulate low-frequency radio interferometric data
**Submitting author:** @JLBLine (<a href="http://orcid.org/0000-0002-9130-5920">Jack L. B. Line</a>)
**Repository:** <a href="https://github.com/JLBLine/WODEN" target ="_blank">https://github.com/JLBLine/WODEN</a>
**Version:** v1.0.0
