### 2021-07-23 #3517 @JLBLine
Roger that @danielskatz, no problem. Looking through the list of potential reviewers, I picked out a number who are both astronomers and have CUDA listed in their languages: devanshkv, Fipanther, adam-m-jcbs, hopefully that helps
