### 2021-09-07 #3676 @mkolopanis
@dfm I wanted to disclose some potential COI I have with this review.

Both Jack Line and myself are collaborators on the Murchison Widefield Array (MWA, the array for which WODEN is specifically designed).
I would like to request this COI be waived since we work independently with no direct communication or collaboration on any papers or projects.

Much of Jack's work focuses on the Real Time System (RTS) processing pipeline.
Meanwhile when I do perform work for the MWA, it is through their other power spectrum/imaging pipeline (FHD+Eppsilon) or through an independent analysis pipeline and packages.
