### 2021-09-07 #3676 @dfm
@mkolopanis: Thanks for this! I am happy to waive this COI since the collaboration is large and if you feel you can provide an impartial review and if @JLBLine is happy with that. Let me know here or via email if you have any concerns about this. Thanks!
