### 2021-09-08 #3676 @JLBLine
@dfm @mkolopanis Yup happy to have the COI waived, makes sense to me
