### 2021-09-09 #3676 @mkolopanis
@JLBLine thanks for this package! WODEN seem like it will be great addition to growing list of simulators for the interferometric world with the use of CUDA acceleration and unique basis function support.

I have some questions and recommendations after my first read through of the paper, and my experience following the examples provided in the documentation.

## Installation
The detailed installation guide found on read the docs is helpful, however, I would recommend rearranging the sections so `How to install Dependencies` directly follows the `Dependencies` section. This could be very useful for users who may be new to or uncomfortable with installing C applications from source. This would require moving the `How to install Dependencies` section to be a sub-section in the Installation section.

The post compilation required section confuses me. I understand the necessity of adding the executables into the path but is it possible to set a custom prefix for compilation to be able to run `make install`  similar to other applications without the need for the `init_WODEN.sh` scritp? I am not too familiar with building C applications. Maybe @plaplant has more thoughts on this.

Another issue I ran into is that while some of these packages are available from apt repositories, attempting to install via apt produced errors but compiling directly from source does not. This may be a product of my environment though. I have been conducting tests using the `nvidia/cuda:11.4.1-devel-ubuntu18.04` docker image and I know some things do not always work as expected in docker without the right configuration.  @JLBLine do you have any insight on this or have you tried installing json-c and erfa via package managers?

## Functionality
The software writes visibilities to disk in the `uvfits` format, however the files written do not appear to fully comply with the fits standard outlined in [AIPS memo 117](http://www.aips.nrao.edu/aipsmemo.html). There are missing mandatory header fields outlined in Table 8. Providing fully compliant uvfits output would make visibility files (and therefore WODEN itself) more modular and accessible from multiple processing pipelines and packages. As an example, the output files are easily converted to CASA measurement sets for imaging in CASA, but can not be read by the [`pyuvdata`](https://github.com/RadioAstronomySoftwareGroup/pyuvdata) python package due to the missing header fields.

It would be encouraging to provide evidence that WODEN's computed visibilities match analytic expectations where possible. The `ctest` suite appears to make comparisons between calculated visibilities and analytically solvable configurations, but never mentions this is done (only that they "check your system / the code runs as expected."). I think it adding to the documentation that tests are performed against analytic visibilities would be helpful.

## Documentation

The link in the docs to ` polarised_source_and_FEE_beam.ipynb which lives here_` appears to redirect to the same page, but viewing the source for the page I can find the actual link and found the accompanying notebook very helpful for how polarization is handled in WODEN.

### statement of need
Regarding the statement of need in the documentation, I mentioned earlier that WODEN would be a great addition to the every growing list of interferometric visibility simulators. But given that the number of simulators is quite large, I think a quantitative comparison would provide more definitive evidence for the need of WODEN over other simulators mentioned especially for the intended use in the MWA. The Line 2020 paper makes quantitative comparisons between WODEN's shapelet approach and other source models in the imaging and power spectrum domain, but these comparisons are useful at showing the high level and end product results. The comparisons were made to evaluate the methods outlined in the paper and not the computational efficiency or accuracy of WODEN compared to other simulators.

### Community Guidelines
The author encourages questions be emailed to him directly on repo's main README and a note is provided in the documentations to create a github issue if a certain beam feature is desied. However, no written community guidelines are provided for researchers looking contribute to the repository. An addition of these guidelines and what steps to undertake would make the process of contributing more accessible to other researchers.

### Automated Tests
No automated test suites exist for this repo, but the author does provide instructions for running `ctest` to validate the installation. Many CI/CD platforms exist where building and testing could be automated.  Setting these up, along with detailed community guidelines, would go a long way to making the code more accessible to future developers to verify/validate that their contributions preserve functionality. I would be happy to help set up CI for this repo if the author is interested. On a related note, I've never made a C application that also tracks code coverage, but I've found a few details about this as well. maybe @plaplant has experience with either of these too?

## Software Paper
There are a few general items in the paper.

- (known in the field as "point sources) -> this parenthetical should be followed by a comma
- Astropy Collaboration citation -> should this be a parenthetical citation?
- Under a formalism like the above -> I understand what is meant by this sentence but it is difficult to parse. Perhaps something along the lines of "Under this discrete sky formalism, upwards of j>=25 x 10^6 or more components..."

### References
The paper refers to pyuvsim and links directly to the repository. I think a reference to the pyuvsim JOSS paper would be fair, but I will leave this  to the discretion of @dfm and @JLBLine  since I am an author on pyuvsim.
