### 2021-09-21 #3676 @mkolopanis
Always nice to close a few easy issues :laughing: @JLBLine if you would like I can try to help with some of this. Would you prefer your repository to be forked for a PR or directly edit on the main repository?

Regarding the statement of need, I agree it feels like a lot of work. I honestly wish we as a community has an organized curated list of simulators. Maybe that partially _does_ exist as part of the `pyuvsim` repository with its reference simulation suite. Let me get back to you on this. This could be as simple as what you suggested and a small PR on `pyuvism` to add WODEN to its comparison documentation.
