### 2021-09-21 #3676 @plaplant
@JLBLine this package is very interesting! Thanks very much for the contribution. In general I find the paper and the code well-written and informative. I have some comments that are similar to those of @mkolopanis, mostly around installation and documentation.

### Installation
I was working as an unprivileged user on a shared cluster (which I imagine is a common-ish use case), and as such was forced to build many of the dependencies by hand. This also meant that my libraries ended up in non-standard locations (e.g., `$HOME/local/json-c/lib`). Though `cmake` was able to find these libraries by editing the values of, e.g., `JSONC_ROOT`, I was still required to update the value of `LD_LIBRARY_PATH` when attempting to run `woden` or `run_woden.py`. If possible, it would be more user-friendly to `export` an appropriate `LD_LIBRARY_PATH` as part of the `init_woden.sh` command. Although there is a section of the installation documentation explaining this, making it more automated would be helpful.

Also, to address the previous point on my end, I had updated the `init_woden.sh` script to update my `LD_LIBRARY_PATH`, but noticed that the file was overwritten upon a subsequent invocation of `cmake`. Admittedly I do not have a ton of experience using `cmake`, but this was counter-intuitive to me. If this behavior is standard, then perhaps a statement in the documentation warning the user to this effect would be helpful.

### Test suite
I greatly appreciate the fact that there are unit tests defined, and building and running them through `ctest` was relatively straightforward. Along the lines of what @mkolopanis suggested, it would be nice to have automated CI running of these tests, though I realize that requires access to a testing provider that has GPUs available to test the full functionality.

Also, it might be good to have a more detailed description of what the specific tests being run do, and what parts of the code they are testing. Getting detailed coverage of individual lines would be ideal, though I understand this may be difficult if such tools do not exist for C/CUDA code. (I personally have not used `ctest` in another context, so I do not know how feasible this is.)

### Documentation
The online documentation on ReadTheDocs is incredibly detailed and very useful for a new user getting familiar with the software package. The pages are well-written and contain helpful mathematical formulas and visualizations when appropriate. The in-line comments on the code are also very helpful for explaining what the code is doing in certain points. Nice work!

### Code Timing
I ran the simulations suggested on [the Fornax A documentation page](https://woden.readthedocs.io/en/latest/examples/fornaxA_sim.html), and found performance in line with those listed on the page. I am using a Tesla V100 card, and the multi-scale CLEAN and shapelet simulations took approximately 15 seconds and 16 seconds, respectively. I think it's a nice feature to have simple benchmarking scripts available easily for the user, which helps demonstrate the code's performance. I don't believe there are any changes to make here, I'm just remarking here for posterity that I see performance similar to what is in the docs.

### Functionality Documentation
Reading the [page on source specification](https://woden.readthedocs.io/en/latest/operating_principles/skymodel.html) and [the one on primary beams](https://woden.readthedocs.io/en/latest/operating_principles/primary_beams.html), it seems (to me) that there is full support for handling polarized sources and correctly propagating them through to output Stokes visibilities. However, this is not explicitly noted in the documentation (and it should be highlighted, because it's a nice feature!). The mathematical expression of the radio interferometer measurement equation (RIME) in the docs and in the paper should include reference to polarization. Also, is there current or future-planned support for outputting instrumental polarizations (e.g., xx and yy) in addition to Stokes parameters? This would be helpful for comparing with other simulation packages like `pyuvsim`.

### Software Paper
The paper is clear and well-written. I have just a few small comments to slightly improve the clarity of some points:
* As mentioned above, a more explicit discussion of polarization (both of sources and the primary beam) should be included, as well as being included in the equation.
* It might be worth mentioning exactly how the pixelization of the sphere is done (e.g., a regular grid in (l, m), HEALPix pixels, something else?), as it affects how things like the sky model and beam interpolation are handled.
* Although there are benchmarking results included in both the reference paper (Line et al. 2020) and some of the documentation pages, it might be nice to include it in the "Example application" section as well, to highlight the speed of the calculation.

I agree with @mkolopanis that an apples-to-apples comparison with other simulation packages would be quite nice, but might be a fair amount of work and is likely beyond the scope of this paper.
