### 2021-09-29 #3676 @mkolopanis
regarding the simulator comparison, I think it will be reasonable to add WODEN to the comparison section of pyuvsim's documents. That can easily be handled outside of this paper.
