### 2021-09-29 #3676 @mkolopanis
It also seems like CircleCI allows for GPU execution on some of their runners according to their [documentation](https://circleci.com/docs/2.0/configuration-reference/). But I've turned up the same nothing regarding coverage for cuda code.
You could also use github actions with a self-hosted runner with a GPU available. I've seen that done in other projects of mine and if you have a dedicated GPU accelerated server with a FQDN this is doable too. There are some security concerns that github identifies about using self-hosted runners on public repos though which should be considered.
