### 2021-09-30 2f44420
Tried to make uvfits file work with `pyuvdata`. Have tested by being able to read a WODEN uvfits file into a `pyuvdata.UVData` object. Involved a number of extra keywords and values, which meant changing the `run_woden.get_LST` function into `run_woden.get_uvfits_date_and_position_constants` function, and modifying the `run_woden.make_antenna_table` and `run_woden.create_uvfits` functions. Have updated the cmake tests to work with the new functionality, and well as a new test to read in a uvfits file using `pyuvdata`. Also fixed up the incorrect hyperlink to `polarised_source_and_FEE_beam.ipyn
[…]
