+++ src/run_woden.py
@@ -4,7 +4,7 @@ the GPU WODEN code. Author: J.L.B. Line
 """
 from __future__ import print_function
 from astropy.io import fits
-from astropy.time import Time
+from astropy.time import Time, TimeDelta
 from astropy.coordinates import EarthLocation
 from astropy import units as u
 
@@ -78,11 +78,16 @@ def calc_jdcal(date):
     ##then the fraction goes into the data array for PTYPE5
     return jd_day, jd_fraction
 
-def get_LST(latitude=None,longitude=None,date=None,height=None):
+def get_uvfits_date_and_position_constants(latitude=None,longitude=None,
+                                           date=None,height=None):
     """
-    For the given Earth location and UTC date return the local sidereal time
-    in degrees. Uses `astropy.time.Time`_ and `astropy.coordinates.EarthLocation`_
-
+    Returns a number of date and time based values that are needed for uvfits
+    headers. For the given Earth location and UTC date return the local sidereal
+    time (deg), the Greenwich sidereal time at 0 hours on the given date (deg),
+    the rotational speed of Earth on the given date (in degrees per day), and
+    the difference between UT1 and UTC.
+    Uses `astropy.time.Time`_ and `astropy.coordinates.EarthLocation`_ to make
+    the calculations.
 
     .. _astropy.time.Time: https://docs.astropy.org/en/stable/time/
     .. _astropy.coordinates.EarthLocation: https://docs.astropy.org/en/stable/api/astropy.coordinates.EarthLocation.html?highlight=EarthLocation
@@ -99,7 +104,14 @@ def get_LST(latitude=None,longitude=None,date=None,height=None):
 
     Returns
     -------
-
+    LST_deg : float
+        Local sidereal time (degrees)
+    GST0_deg : float
+        Greenwich sidereal time at 0 hours on the given date (degrees)
+    DEGPDY : float
+        Rotational speed of Earth on the given date (degrees per day)
+    ut1utc : float
+        Difference between UT1 and UTC (secs)
     """
     ##Setup location
     observing_location = EarthLocation(lat=latitude*u.deg, lon=longitude*u.deg, height=height)
@@ -107,7 +119,29 @@ def get_LST(latitude=None,longitude=None,date=None,height=None):
     observing_time = Time(date, scale='utc', location=observing_location)
     ##Grab the LST
     LST = observing_time.sidereal_time('apparent')
-    return LST.value*15.0
+    LST_deg = LST.value*15.0
+
+    ##uvfits file needs to know the greenwich sidereal time at 0 hours
+    ##on the date in question
+    zero_date = date.split('T')[0] + "T00:00:00"
+    zero_time = Time(zero_date, scale='utc', location=observing_location)
+    GST0 = zero_time.sidereal_time('apparent', 'greenwich')
+    GST0_deg = GST0.value*15.0
+
+    ##It also needs to know the rotational rate of the Earth on that day, in
+    ##units of degrees per day
+    ##Do this by measuring the LST exactly a day later
+
+    date_plus_one_day =  observing_time + TimeDelta(1*u.day)
+    LST_plusone = date_plus_one_day.sidereal_time('apparent')
+
+    LST_plusone_deg = LST_plusone.value*15.0
+
+    DEGPDY = 360.0 + (LST_plusone_deg - LST_deg)
+
+    ut1utc = float(observing_time.delta_ut1_utc)
+
+    return LST_deg, GST0_deg, DEGPDY, ut1utc
 
 def RTS_encode_baseline(b1, b2):
[…]
