+++ docs/sphinx/API_reference/python_code/run_woden.rst
@@ -4,6 +4,8 @@
 This is the main way to control the ``WODEN`` executable, and ensure good
 arguments are supplied. I suggest you only run ``WODEN`` through this script.
 
+.. _run_woden command line running options:
+
 *Command line running options*
 -------------------------------
 
