+++ docs/sphinx/API_reference/python_code/uv2ms.rst
@@ -3,6 +3,8 @@
 
 Helper script to convert ``uvfits`` files to ``measurement sets`` using ``casa``.
 
+.. _uv2ms command line running options:
+
 *Command line running options*
 -------------------------------
 
