### 2021-09-30 #3676 @JLBLine
Thanks for checking that out @mkolopanis  - I've just looked into CircleCI which seemed a great solution, until I looked into signing up, and you can only get GPUs on a 'custom pricing plan' so it won't be free. I'm reluctant to use a self-hosted runner, mostly because I don't want random people spinning up my desktop GPU when I might be using it, and the only other GPUs I have access to are queue-based super clusters. I'm not trying to be miserly, but I don't really want to pay to host a cloud-based option, because whatever research funding I have will always be temporary given the nature of academia.

For the comparison with other simulators, that sounds good. Steven Murray mentioned something similar about getting a `pyradiosky` model through WODEN as well, to compare to the Hera simulators. Maybe once we've got this review finished we can have a meeting to work out how to tackle this all in one foul swoop.

I've just pushed a new branch called `joss_review`, which I think should write uvfits files that work with `pyuvdata`. At least they work to the extent that you can create a `pyuvdata.UVData` object and read in a WODEN uvfits file. Please have a quick look and let me know if it works as you would expect.

If you are keen to do anything to WODEN in the mean time, I'd prefer a new fork (or branch) and PR rather than working on master.
