### 2021-09-30 #3676 @JLBLine
Thanks for the review @plaplant. Please see some responses below. Once I've finished this response, I'll try and grab all the relevant points from both your and @mkolopanis reviews and turn them into git issues. I should then be able to notify you guys once I think I've solved each of them and how, and maybe get you guys to verify them so I can close them? If you see any other issues I've missed feel free to add them.

## Installation
_I was working as an unprivileged user on a shared cluster (which I imagine is a common-ish use case), and as such was forced to build many of the dependencies by hand. This also meant that my libraries ended up in non-standard locations (e.g., $HOME/local/json-c/lib). Though cmake was able to find these libraries by editing the values of, e.g., JSONC_ROOT, I was still required to update the value of LD_LIBRARY_PATH when attempting to run woden or run_woden.py. If possible, it would be more user-friendly to export an appropriate LD_LIBRARY_PATH as part of the init_woden.sh command. Although there is a section of the installation documentation explaining this, making it more automated would be helpful._

Yeah I think I can see a way for this to happen, I'll have a go.

_Also, to address the previous point on my end, I had updated the init_woden.sh script to update my LD_LIBRARY_PATH, but noticed that the file was overwritten upon a subsequent invocation of cmake. Admittedly I do not have a ton of experience using cmake, but this was counter-intuitive to me. If this behavior is standard, then perhaps a statement in the documentation warning the user to this effect would be helpful._

So the `init_woden.sh` file actually lives in the `WODEN/templates` folder, and is copied into the `build` directory upon compilation. I don't think this is standard as I don't think having a bash file that you call for setup is standard at all, it just made sense to me (this is the first C/CUDA/CMake project I've ever written so I am admittedly flying by the seat of my pants). What would be more standard would be to setup a `make install` option as @mkolopanis suggested. I think I'll do that, as well as keeping the `init_woden.sh` as an option if you don't want to `make install`, and leave the warning to the user that it will be overwritten, and can be edited in the `WODEN/templates` folder if required.

## Test suite
_I greatly appreciate the fact that there are unit tests defined, and building and running them through ctest was relatively straightforward. Along the lines of what @mkolopanis suggested, it would be nice to have automated CI running of these tests, though I realize that requires access to a testing provider that has GPUs available to test the full functionality._

As I've been discussing with @mkolopanis, this seems impossible to do for free, and I don't really want to have to pay to support online testing.

_Also, it might be good to have a more detailed description of what the specific tests being run do, and what parts of the code they are testing. Getting detailed coverage of individual lines would be ideal, though I understand this may be difficult if such tools do not exist for C/CUDA code. (I personally have not used ctest in another context, so I do not know how feasible this is.)_

Yeah the joys of CUDA rearing it's head again. I can certainly put more documentation into what the tests are doing and why, which I'll work on. I _think_ I might be able to hook up CMake-codecov ( https://github.com/RWTH-HPC/CMake-codecov ) to my python and C tests to get coverage for the python/C code, so I'll look into that. Pretty sure it won't work with the CUDA code though.

## Documentation
_The online documentation on ReadTheDocs is incredibly detailed and very useful for a new user getting familiar with the software package. The pages are well-written and contain helpful mathematical formulas and visualizations when appropriate. The in-line comments on the code are also very helpful for explaining what the code is doing in certain points. Nice work!_

Sweet, thanks!

## Code Timing
_I ran the simulations suggested on the Fornax A documentation page, and found performance in line with those listed on the page. I am using a Tesla V100 card, and the multi-scale CLEAN and shapelet simulations took approximately 15 seconds and 16 seconds, respectively. I think it's a nice feature to have simple benchmarking scripts available easily for the user, which helps demonstrate the code's performance. I don't believe there are any changes to make here, I'm just remarking here for posterity that I see performance similar to what is in the docs._

Cool, good to know. I think you'll find that for these small simulations, large amounts of time are actually I/O, especially because I've been caveman in the way I write out a binary file from C/CUDA and read it into python to write the uvfits, rather than attempting to write uvfits directly from with C. I found with the V100, for 200,000 point sources, 16 time steps and 384 freq channels, the V100 ran three times faster out of the box than the 1080 Ti. Tasty!

## Functionality Documentation
_Reading the page on source specification and the one on primary beams, it seems (to me) that there is full support for handling polarized sources and correctly propagating them through to output Stokes visibilities. However, this is not explicitly noted in the documentation (and it should be highlighted, because it's a nice feature!). The mathematical expression of the radio interferometer measurement equation (RIME) in the docs and in the paper should include reference to polarization. Also, is there current or future-planned support for outputting instrumental polarizations (e.g., xx and yy) in addition to Stokes parameters? This would be helpful for comparing with other simulation packages like pyuvsim._

Good point about the polarisation, I'll update that in the docs and the paper. Right now there are no plans for outputting the instrumental pols, partly because I haven't researched the maths of how to change Stokes input params into instrumental pols yet.

## Software Paper
_The paper is clear and well-written. I have just a few small comments to slightly improve the clarity of some points:_

_As mentioned above, a more explicit discussion of polarization (both of sources and the primary beam) should be included, as well as being included in the equation.
It might be worth mentioning exactly how the pixelization of the sphere is done (e.g., a regular grid in (l, m), HEALPix pixels, something else?), as it affects how things like the sky model and beam interpolation are handled.
Although there are benchmarking results included in both the reference paper (Line et al. 2020) and some of the documentation pages, it might be nice to include it in the "Example application" section as well, to highlight the speed of the calculation._

I shall make changes to address these points. Regarding the first point, there is _no_ pixelisation done internally to WODEN. The RIME equation is just done for every direction in the sky model, for all frequencies and time steps. All pixelisation happens in the sky model (made of point, Gaussian, and shapelet models), so you could put in any pixelisation scheme you want.
