### 2021-09-30 JLBLine/WODEN#15 @JLBLine
JOSS review issue 2 (Installation)
Make a "make install" option via CMake

https://github.com/openjournals/joss-reviews/issues/3676
