### 2021-09-30 JLBLine/WODEN#16 @JLBLine
JOSS review issue 3 (Installation)
Export a user-specific version of `init_woden.sh` that includes a custom LD_LIBRARY_PATH for systems with non-standard installations. Warn the user in the Documentation that `init_woden.sh` is overwritten in the `build` directory every time. I think this can be achieved by searching `CMakeCache.txt` and checking for any non-standard paths.

https://github.com/openjournals/joss-reviews/issues/3676
