### 2021-09-30 JLBLine/WODEN#17 @JLBLine
JOSS review issue 4 (Tests)
Detail what the unit / integration tests are actually doing in the Documentation, as a standalone section inside `Testing via ctest`. Add plenty of comments and explanations inside the tests themselves as well.

https://github.com/openjournals/joss-reviews/issues/3676
