### 2021-09-30 JLBLine/WODEN#18 @JLBLine
JOSS review issue 5 (Code coverage)
Investigate the use of CMake-codecov to get test coverage of the python/C code

https://github.com/openjournals/joss-reviews/issues/3676
