### 2021-09-30 JLBLine/WODEN#19 @JLBLine
JOSS review issue 6 (uvfits files)
Make the uvfits files work with pyuvdata

https://github.com/openjournals/joss-reviews/issues/3676
