### 2021-09-30 JLBLine/WODEN#20 @JLBLine
JOSS review issue 7 (Community Guidelines)
Add some community guidelines for how to contribute

https://github.com/openjournals/joss-reviews/issues/3676
