### 2021-09-30 JLBLine/WODEN#21 @JLBLine
JOSS review issue 8 (Software paper)
Make edits in response to the following comments:

- (known in the field as "point sources) -> this parenthetical should be followed by a comma.
- Astropy Collaboration citation -> should this be a parenthetical citation?
- Under a formalism like the above -> I understand what is meant by this sentence but it is difficult to parse. Perhaps something along the lines of "Under this discrete sky formalism, upwards of j>=25 x 10^6 or more components…"
- The mathematical expression of the radio interferometer measurement equation (RIME) in the docs and in the paper should include reference to polarization.
- It might be worth mentioning exactly how the pixelization of the sphere is done (e.g., a regular grid in (l, m), HEALPix pixels, something else?), as it affects how things like the sky model and beam interpolation are handled.
- Although there are benchmarking results included in both the reference paper (Line et al. 2020) and some of the documentation pages, it might be nice to include it in the "Example application" section as well, to highlight the speed of the calculation.

https://github.com/openjournals/joss-reviews/issues/3676
