### 2021-09-30 #3676 @dfm
> Thanks for checking that out @mkolopanis - I've just looked into CircleCI which seemed a great solution, until I looked into signing up, and you can only get GPUs on a 'custom pricing plan' so it won't be free. I'm reluctant to use a self-hosted runner, mostly because I don't want random people spinning up my desktop GPU when I might be using it, and the only other GPUs I have access to are queue-based super clusters. I'm not trying to be miserly, but I don't really want to pay to host a cloud-based option, because whatever research funding I have will always be temporary given the nature of academia.

@JLBLine: We definitely don't require something like this for a JOSS review as long as the reviewers can execute the code on their own machines, so I think we're ok! It would be great if some of the free CI options started supporting GPU testing someday, but we're definitely not there yet.
