### 2021-10-01 620bd12
Initial setup at getting code coverage working for python/C code. The cmake testing code now uses the git submodule CMake-codecov, which generates C code coverage data. A new script, `coverage_outputs/create_cov_reports.sh` can be run after running `ctest`, which will covert the `C` coverage output using `gcov`, and run the `python` tests using `coverage`. All these outputs can then be pushed to "codecov" using the script `coverage_outputs/send_reports_to_codecov.sh`.
