### 2021-10-01 JLBLine/WODEN#18 @JLBLine
Ok, with this commit https://github.com/JLBLine/WODEN/commit/620bd121862deca2468659a27b8476756ec13ddc, I've added a way to locally generate code coverage, that can be pushed to "codecov" using a specific token. It relies on running the local CMake tests to generate the coverage outputs. I plan on putting instructions on how obtain this token in the "Community Guidelines". I've written a README in the new directory `coverage_outputs` on how to generate the coverage reports and send them off to "codecov".

To fully automate everything upon pushing, I'd have to separate out all CUDA calls from the CMake setup and have a some kind of way of switching off CUDA entirely from CMake for the code coverage generation. Might be possible but will take quite some time to develop methinks. At least this way, I can give this token to anyone trusted enough to be a contributor, and they can update the "codecov" coverage report themselves before pushing changes.
