### 2021-10-01 JLBLine/WODEN#19 @JLBLine
Right, with this commit 2f44420ce5c8987cf2f964d4b797d326c3459fb8, I believe I've made my `uvfits` files compatible with `pyuvdata`, in so far as I can make a `pyuvdata.UVData` object and read my `uvfits` files in with no errors.
