### 2021-10-08 c258a1f
Updated the documentation to respond to JOSS review issue 1
