### 2021-10-08 184eff7
Setup up "sudo make install" option for CMake, that will install the `woden` executable in the default 'GNUInstallDirs' as defined by CMake - for me that was /usr/local/bin for the `woden` and `run_woden.py` executables and /use/locl/lib for the `libwodenCUDA.so` and `libwodenC.so` libraries. Meant modifying the way `run_woden.py` looks for the `woden` executable, as well as the imaging scripts inside of the `examples` dir. Still need to update the documentation to reflect new installation method
