### 2021-10-08 8326e24
Added some words to the documentation about the new option to run `make install`, and made adding the `init_WODEN.sh` script to your bashrc optional instead of required.
