### 2021-10-11 JLBLine/WODEN#19 @JLBLine
@plaplant @mkolopanis I believe this has been addressed, please have a look. I'm not familiar with `pyuvdata` so beyond reading data in I'm not sure what tests to run
