### 2021-10-11 74f71f9
Added the new script `update_init_WODEN.py` that reads `CMakeCache.txt` to find library paths. Then adds lines to `init_WODEN.sh` to add those paths to LD_LIBRARY_PATH so `woden` can find them at run time. Updated the documentation accordingly.
