+++ docs/sphinx/installation/installation.rst
@@ -129,7 +129,17 @@ If you don't run ``make install``, ``run_woden.py`` won't be able to find the ``
 
   source /path/to/your/location/WODEN/build/init_WODEN.sh
 
-to your ``~/.bash_rc`` (where you replace ``/path/to/your/location`` to wherever you installed ``WODEN``). This will create the variable ``$WODEN_DIR``, and add it to your ``$PATH``. This allows ``run_woden.py`` to find the ``woden`` executable.
+to your ``~/.bash_rc`` (where you replace ``/path/to/your/location`` to wherever you installed ``WODEN``). This will create the variable ``$WODEN_DIR``, and add it to your ``$PATH``. Furthermore, ``init_WODEN.sh`` is generated by the script ``src/update_init_WODEN.py``, which looks through ``CMakeCache.txt`` for the locations of ``ERFA``, ``HDF5``, ``JSONC``, ``PAL``. It then appends lines to ``init_WODEN.sh`` to add these locations to ``LD_LIBRARY_PATH``, so ``woden`` can find these libraries at run time. For example, on my machine, ``init_WODEN.sh`` ends up looking like::
+
+  ##This line finds the current directory at sets the env variable WODEN_DIR
+  export WODEN_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+  ##This adds the line to PATH
+  export PATH=$WODEN_DIR:$PATH
+  ##Add library paths to LD_LIBRARY_PATH so the can be found at runtime
+  export LD_LIBRARY_PATH=/usr/lib/x86_64-linux-gnu/hdf5/serial/:$LD_LIBRARY_PATH
+  export LD_LIBRARY_PATH=/usr/local/lib/:$LD_LIBRARY_PATH
+
+.. note:: Every time you run ``make``, ``init_WODEN.sh`` is regenerated, so any edits you make will be overwritten. I suggest any other customisation of you ``LD_LIBRARY_PATH`` happens in your ``~/.bashrc`` or equivalent.
 
 Post compilation (optional)
 ###############################
