### 2021-10-11 JLBLine/WODEN#16 @JLBLine
@plaplant @mkolopanis With commit 74f71f95b34df3b91dcd853f72c49492f2e5920b, I've added new code that generates `init_WODEN.sh` with lines that append the locations of the `erfa`, `pal`, `json`, and `hdf5` libraries to the `LD_LIBRARY_PATH`. This means `woden` should be able to find them at run time. Added a note to the documentation that this is overwritten everytime `make` is run
