### 2021-10-11 c29d7ee
Added some information to `cmake_testing.rst` to instruct how to initiate the `git submodule`, which is now necessary to run the `ctest` and gather code coverage
