### 2021-10-11 JLBLine/WODEN#18 @JLBLine
Ah oops to run the `ctest` code now you'll need to follow the instructions I've added in commit c29d7eeaf608cdd9da9381289e779d854b752458 to the "Testing via `ctest`" documentation section about initiating a `git submodule`. Otherwise the CMake code can't find what it needs when you run `cmake .. -DTARGET_GROUP=test`
