+++ docs/sphinx/testing/cmake_testing.rst
@@ -80,3 +80,34 @@ You should see something like the following if successful::
 If you want more detail of what the tests are doing, run::
 
   $ ctest --verbose
+
+What do the tests actually do?
+---------------------------------
+
+The tests are located in ``WODEN/cmake_testing``, and each directory within contains tests
+for a different file from ``WODEN/src``. Within each test directory, there are separate files for testing different functions, which include the function name. As an example, the directory ``WODEN/cmake_testing/array_layout`` contains tests for the file ``WODEN/src/array_layout.c``, and contains test files that test the following functions::
+
+  cmake_testing/array_layout/test_calc_XYZ_diffs.c -> src/array_layout.c::calc_XYZ_diffs
+  cmake_testing/array_layout/test_RTS_ENH2XYZ_local.c -> src/array_layout.c::RTS_ENH2XYZ_local
+  cmake_testing/array_layout/test_RTS_PrecessXYZtoJ2000.c -> src/array_layout.c::RTS_PrecessXYZtoJ2000
+
+.. note:: For those unfamiliar with ``CMake`` testing, even though the tests are located in ``WODEN/cmake_testing/``, when you run ``ctest``, the test files are copied and run in ``WODEN/build/cmake_testing``, so any output from the tests will be located there.
+
+The sections below give an outline of the tests performed in each directory.
+
+``C`` code tests:
+
+.. toctree::
+   :maxdepth: 1
+
+   cmake_testing/array_layout
+
+
+``CUDA`` code tests:
+
+.. toctree::
+   :maxdepth: 1
+
+   cmake_testing/calculate_visibilities
+
+.. note:: To be able to test ``CUDA`` functions that are designed to work solely in GPU memory, it's necessary to write wrapper functions that allocate GPU memory, pass the data into the ``CUDA`` code to be tested, and then copy the results back into host memory. I've kept these 'intermediate' test functions inside the ``*.cu`` files that contain the code being tested, as it's not straight forward / performance degrading to have them in separate files. On casual inspection it looks like there are many functions in the ``*.cu`` files I haven't written tests for, but the extra functions are there *because* of testing. Sigh.
