### 2021-10-20 c26cde9
Plenty of progress through documenting the tests. Have added a notebook `cmake_testing/source_components/check_accuracy_of_WODEN.ipynb` with my initial experiment on quantifying the absolute accuracy of the measurement equation in WODEN.
