+++ cmake_testing/source_components/test_kern_calc_visi_point.c
@@ -80,8 +80,9 @@ void test_kern_calc_visi_point_Varylmn(int beamtype) {
       frequency = freq_base + freq_step*freq_inc;
       wavelength = VELC / frequency;
       for (int baseline = 0; baseline < num_baselines; baseline++) {
-        args_ft->us[count] = ((baseline + 1)*10) / wavelength;
-        args_ft->vs[count] = ((baseline + 1)*10) / wavelength;
+        args_ft->us[count] = ((baseline + 1)*100) / wavelength;
+        args_ft->vs[count] = ((baseline + 1)*100) / wavelength;
+        //ws are usually smaller than u,v
         args_ft->ws[count] = ((baseline + 1)*10) / wavelength;
 
         args_ft->allsteps_wavelengths[count] = wavelength;
@@ -145,7 +146,8 @@ void test_kern_calc_visi_point_VarylmnNoBeam(void) {
 /*
 Test the __device__ code that updates the summed visibilities by grabbing the
 correct beam gain and mesurement equation, multiplying and summing onto the visi
-Here we keep the component visibilities and fluxes constant and vary the beam gains
+Here we test for multiple l,m,n, vary the flux of the components, but keep
+the beam gains constant
 Test works for all primary beam types
 */
 void test_kern_calc_visi_point_VarylmnVaryFlux(int beamtype) {
@@ -178,9 +180,9 @@ void test_kern_calc_visi_point_VarylmnVaryFlux(int beamtype) {
     args_ft->primay_beam_J11[visi] = 1.0 + I*0.0;
   }
 
-  //Just stick Stokes I to 1.0, SI to zero, and reference freqs to 150MHz
+  //Just stick Stokes I to the component value, and SI to -0.8
   for (size_t comp = 0; comp < num_components; comp++) {
-    args_ft->flux_I[comp] = comp;
+    args_ft->flux_I[comp] = comp + 1;
     args_ft->SIs[comp] = -0.8;
     args_ft->component_freqs[comp] = 150e+6;
   }
@@ -200,8 +202,9 @@ void test_kern_calc_visi_point_VarylmnVaryFlux(int beamtype) {
       wavelength = VELC / frequency;
 
       for (int baseline = 0; baseline < num_baselines; baseline++) {
-        args_ft->us[count] = ((baseline + 1)*10) / wavelength;
-        args_ft->vs[count] = ((baseline + 1)*10) / wavelength;
+        args_ft->us[count] = ((baseline + 1)*100) / wavelength;
+        args_ft->vs[count] = ((baseline + 1)*100) / wavelength;
+        //ws are usually smaller than u,v
         args_ft->ws[count] = ((baseline + 1)*10) / wavelength;
 
         args_ft->allsteps_wavelengths[count] = wavelength;
@@ -227,7 +230,7 @@ void test_kern_calc_visi_point_VarylmnVaryFlux(int beamtype) {
           args_ft->primay_beam_J10, args_ft->primay_beam_J11);
 
   //Check all results are within 0.1% of expected value
-  float frac_tol = 5e-3;
+  float frac_tol = 1e-3;
   test_visi_outputs(num_visis, num_components, num_baselines, num_freqs,
                     frac_tol, beamtype, args_ft, POINT);
 
@@ -262,12 +265,17 @@ void test_kern_calc_visi_point_VarylmnVaryFluxNoBeam(void) {
   test_kern_calc_visi_point_VarylmnVaryFlux(NO_BEAM);
 }
 
-
+/*
[…]
