### 2021-10-20 9c35661
Didn't add changes to one of the c files that tests the measurement equation
