+++ cmake_testing/source_components/test_calc_measurement_equation.c
@@ -9,6 +9,7 @@ void setUp (void) {} /* Is run before every test, put unit init calls here. */
 void tearDown (void) {} /* Is run after every test, put unit clean-up calls here. */
 
 void sincosf(float x, float *sin, float *cos);
+void sincos(double x, double *sin, double *cos);
 
 //External CUDA code we're linking in
 extern void test_kern_calc_measurement_equation(int num_components,
@@ -20,9 +21,11 @@ extern void test_kern_calc_measurement_equation(int num_components,
 #define UNITY_INCLUDE_FLOAT
 
 /*
-Test that the linear SI flux extrapolation code works correctly
+Setup 1000 u,v,w coords and 100 l,m,n coords, and send them off to
+kern_calc_measurement_equation to calculate the measurement equation.
+Check against a calculation made with C
 */
-void test_kern_calc_measurement_equation_GiveCorrectValues(void) {
+void test_kern_calc_measurement_equation_ComparedToC(void) {
 
   //Set up some test condition inputs
   int num_baselines = 1000;
@@ -33,9 +36,13 @@ void test_kern_calc_measurement_equation_GiveCorrectValues(void) {
   float max_v = 1000.0;
   float max_w = 20.0;
 
-  float u_inc = (2*max_u) / num_baselines;
-  float v_inc = (2*max_v) / num_baselines;
-  float w_inc = (2*max_w) / num_baselines;
+  // float u_inc = (2*max_u) / num_baselines;
+  // float v_inc = (2*max_v) / num_baselines;
+  // float w_inc = (2*max_w) / num_baselines;
+
+  float u_inc = 2;
+  float v_inc = 2;
+  float w_inc = 0.04;
 
   float *us = malloc(num_baselines*sizeof(float));
   float *vs = malloc(num_baselines*sizeof(float));
@@ -51,8 +58,10 @@ void test_kern_calc_measurement_equation_GiveCorrectValues(void) {
   float *ms = malloc(num_components*sizeof(float));
   float *ns = malloc(num_components*sizeof(float));
 
-  float l_inc = 1.4 / num_components;
-  float m_inc = 1.4 / num_components;
+  // float l_inc = 1.4 / num_components;
+  // float m_inc = 1.4 / num_components;
+  float l_inc = 0.014;
+  float m_inc = 0.014;
   float l, m;
 
   for (size_t component = 0; component < num_components; component++) {
@@ -72,21 +81,22 @@ void test_kern_calc_measurement_equation_GiveCorrectValues(void) {
                                       us, vs, ws, ls, ms, ns, visis);
 
   //Make some expected value arrays
-  float *expec_res = malloc(num_baselines*num_components*sizeof(float));
-  float *expec_ims = malloc(num_baselines*num_components*sizeof(float));
+  double *expec_res = malloc(num_baselines*num_components*sizeof(double));
+  double *expec_ims = malloc(num_baselines*num_components*sizeof(double));
 
   float *visi_res = malloc(num_baselines*num_components*sizeof(float));
   float *visi_ims = malloc(num_baselines*num_components*sizeof(float));
 
   //Fill values with what should have been found
   int ind = 0;
-  float temp, expec_re, expec_im;
+  float temp;
+  double expec_re, expec_im;
 
   for (size_t baseline = 0; baseline < num_baselines; baseline++) {
[…]
