### 2021-10-21 5042bce
Add an external double precision version of the measurement equation code as a comparison
