+++ cmake_testing/source_components/source_components_double.cu
@@ -1,0 +1,120 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include <cuComplex.h>
+#include <complex.h>
+#include <math.h>
+#include "cudacomplex.h"
+// #include "fundamental_coords.h"
+// #include "constants.h"
+// #include "shapelet_basis.h"
+// #include "source_components.h"
+#include "cudacheck.h"
+// #include "woden_struct_defs.h"
+// #include "primary_beam_cuda.h"
+// #include "FEE_primary_beam_cuda.h"
+
+__device__  cuDoubleComplex calc_measurement_equation(double *d_us,
+           double *d_vs, double *d_ws, double *d_ls, double *d_ms, double *d_ns,
+           const int iBaseline, const int iComponent){
+
+  double u, v, w;
+  double l, m, n;
+
+  u = d_us[iBaseline];
+  v = d_vs[iBaseline];
+  w = d_ws[iBaseline];
+
+  l = d_ls[iComponent];
+  m = d_ms[iComponent];
+  n = d_ns[iComponent];
+
+  cuDoubleComplex visi;
+
+  //Not sure why, but get match with OSKAR/RTS sims, and correct location
+  //on sky through WSClean, without negative infront on 2pi
+  double temp = 2*M_PI*( u*l + v*m + w*(n-1) );
+  // sincosf(temp, &(visi.y), &(visi.x));
+
+  visi.y = sin(temp);
+  visi.x = cos(temp);
+
+  return visi;
+}
+
+
+
+__global__ void kern_calc_measurement_equation(int num_components, int num_baselines,
+                  double *d_us, double *d_vs, double *d_ws,
+                  double *d_ls, double *d_ms, double *d_ns,
+                  cuDoubleComplex *d_visis) {
+
+  // Start by computing which baseline we're going to do
+  const int iBaseline = threadIdx.x + (blockDim.x*blockIdx.x);
+  const int iComponent = threadIdx.y + (blockDim.y*blockIdx.y);
+  // if(iBaseline < num_visis && iComponent < num_points) {
+  if(iComponent < num_components && iBaseline < num_baselines) {
+
+    cuDoubleComplex visi;
+    visi = calc_measurement_equation(d_us, d_vs, d_ws, d_ls, d_ms, d_ns,
+                                     iBaseline, iComponent);
+
+    int visi_ind = num_components*iBaseline + iComponent;
+    d_visis[visi_ind] = visi;
+
+  }
+}
+
+extern "C" void test_kern_calc_measurement_equation(int num_components,
+                  int num_baselines,
+                  double *us, double *vs, double *ws,
+                  double *ls, double *ms, double *ns,
+                  double _Complex *visis){
+
+  double *d_us = NULL;
+  double *d_vs = NULL;
+  double *d_ws = NULL;
+  cudaErrorCheckCall( cudaMalloc( (void**)&d_us, num_baselines*sizeof(double) ));
+  cudaErrorCheckCall( cudaMalloc( (void**)&d_vs, num_baselines*sizeof(double) ));
+  cudaErrorCheckCall( cudaMalloc( (void**)&d_ws, num_baselines*sizeof(double) ));
+  cudaErrorCheckCall( cudaMemcpy(d_us, us, num_baselines*sizeof(double), cudaMemcpyHostToDevice ));
+  cudaErrorCheckCall( cudaMemcpy(d_vs, vs, num_baselines*sizeof(double), cudaMemcpyHostToDevice ));
[…]
