### 2021-10-21 0f0c3e1
Didn't add the C code that runs the CUDA double precision test
