+++ cmake_testing/source_components/test_calc_measurement_equation_double.c
@@ -1,0 +1,336 @@
+#include <math.h>
+#include <unity.h>
+#include <stdlib.h>
+#include <complex.h>
+
+#include "constants.h"
+
+void setUp (void) {} /* Is run before every test, put unit init calls here. */
+void tearDown (void) {} /* Is run after every test, put unit clean-up calls here. */
+
+void sincosf(float x, float *sin, float *cos);
+void sincos(double x, double *sin, double *cos);
+
+//External CUDA code we're linking in
+extern void test_kern_calc_measurement_equation(int num_components,
+                  int num_baselines,
+                  double *us, double *vs, double *ws,
+                  double *ls, double *ms, double *ns,
+                  double _Complex *visis);
+
+#define UNITY_INCLUDE_FLOAT
+
+// /*
+// Setup 1000 u,v,w coords and 100 l,m,n coords, and send them off to
+// kern_calc_measurement_equation to calculate the measurement equation.
+// Check against a calculation made with C
+// */
+// void test_kern_calc_measurement_equation_ComparedToC(void) {
+//
+//   //Set up some test condition inputs
+//   int num_baselines = 1000;
+//   int num_components = 100;
+//
+//   //This is a reasonable range based on MWA
+//   double max_u = 1000.0;
+//   double max_v = 1000.0;
+//   double max_w = 20.0;
+//
+//   // double u_inc = (2*max_u) / num_baselines;
+//   // double v_inc = (2*max_v) / num_baselines;
+//   // double w_inc = (2*max_w) / num_baselines;
+//
+//   double u_inc = 2;
+//   double v_inc = 2;
+//   double w_inc = 0.04;
+//
+//   double *us = malloc(num_baselines*sizeof(double));
+//   double *vs = malloc(num_baselines*sizeof(double));
+//   double *ws = malloc(num_baselines*sizeof(double));
+//
+//   for (size_t baseline = 0; baseline < num_baselines; baseline++) {
+//     us[baseline] = -max_u + u_inc*baseline;
+//     vs[baseline] = -max_v + v_inc*baseline;
+//     ws[baseline] = -max_w + w_inc*baseline;
+//   }
+//
+//   double *ls = malloc(num_components*sizeof(double));
+//   double *ms = malloc(num_components*sizeof(double));
+//   double *ns = malloc(num_components*sizeof(double));
+//
+//   // double l_inc = 1.4 / num_components;
+//   // double m_inc = 1.4 / num_components;
+//   double l_inc = 0.014;
+//   double m_inc = 0.014;
+//   double l, m;
+//
+//   for (size_t component = 0; component < num_components; component++) {
+//     l =  -0.7 + l_inc*component;
+//     m =  -0.7 + m_inc*component;
+//     ls[component] = l;
+//     ms[component] = m;
+//     ns[component] = sqrt(1 - l*l - m*m);
+//   }
+//
+//
+//   //Space for outputs
+//   double _Complex *visis = malloc(num_baselines*num_components*sizeof(double _Complex));
+//
+//   //Run the CUDA code
+//   test_kern_calc_measurement_equation(num_components, num_baselines,
+//                                       us, vs, ws, ls, ms, ns, visis);
+//
+//   //Make some expected value arrays
+//   double *expec_res = malloc(num_baselines*num_components*sizeof(double));
[…]
