### 2021-10-22 ca76960
Added an option to switch of precession, useful for testing.
