### 2021-10-26 35cc646
After testing the absolute accuracy of WODEN, I've upgraded the l,m,n coordinates from float to double precision. This involved changing every part of the code touched by the l,m,n, including the test code. Added preliminary materials on testing the effects of the accuracy of WODEN on caliberation. Fixed the imaging scripts in `test_installation` to work without `$WODEN_DIR` being defined.
