+++ src/calculate_visibilities.cu
@@ -25,22 +25,6 @@ extern "C" void calculate_visibilities(array_layout_t *array_layout,
   const int num_freqs = woden_settings->num_freqs;
 
   //Setup chunk_visibility_set to hold the visibility outputs of each
-  //sky chunk simulation
-  // visibility_set_t *chunk_visibility_set = (visibility_set_t *)malloc(sizeof(visibility_set_t));
-  //
-  // chunk_visibility_set->us_metres = (float *)malloc( num_visis * sizeof(float) );
-  // chunk_visibility_set->vs_metres = (float *)malloc( num_visis * sizeof(float) );
-  // chunk_visibility_set->ws_metres = (float *)malloc( num_visis * sizeof(float) );
-  //
-  // chunk_visibility_set->sum_visi_XX_real = (float *)malloc( num_visis * sizeof(float) );
-  // chunk_visibility_set->sum_visi_XX_imag = (float *)malloc( num_visis * sizeof(float) );
-  // chunk_visibility_set->sum_visi_XY_real = (float *)malloc( num_visis * sizeof(float) );
-  // chunk_visibility_set->sum_visi_XY_imag = (float *)malloc( num_visis * sizeof(float) );
-  // chunk_visibility_set->sum_visi_YX_real = (float *)malloc( num_visis * sizeof(float) );
-  // chunk_visibility_set->sum_visi_YX_imag = (float *)malloc( num_visis * sizeof(float) );
-  // chunk_visibility_set->sum_visi_YY_real = (float *)malloc( num_visis * sizeof(float) );
-  // chunk_visibility_set->sum_visi_YY_imag = (float *)malloc( num_visis * sizeof(float) );
-
   visibility_set_t *chunk_visibility_set = setup_visibility_set(num_visis);
 
   //Copy across settings for simulation - we always simulate all time steps
@@ -132,14 +116,7 @@ extern "C" void calculate_visibilities(array_layout_t *array_layout,
                         cudaMemcpyHostToDevice ));
   }
 
-  // RTS_MWA_FEE_beam_t *FEE_beam;
-  // RTS_MWA_FEE_beam_t *FEE_beam_zenith;
-
   if (beam_settings->beamtype == FEE_BEAM){
-
-    // beam_settings = cropped_sky_models->beam_settings[0];
-    // FEE_beam = beam_settings.FEE_beam;
-    // FEE_beam_zenith = beam_settings.FEE_beam_zenith;
 
     //Only needs to be done once per frequency as zenith never moves, so do
     //this outside the sky model chunk loop
@@ -160,9 +137,6 @@ extern "C" void calculate_visibilities(array_layout_t *array_layout,
 
     catsource_t catsource;
     catsource = cropped_sky_models->catsources[chunk];
-
-    // beam_settings_t beam_settings;
-    // beam_settings = cropped_sky_models->beam_settings[chunk];
 
     //Make sure the temp visis are 0 at the start of each chunk
     for (int visi = 0; visi < num_visis; visi++) {
@@ -235,15 +209,15 @@ extern "C" void calculate_visibilities(array_layout_t *array_layout,
     cuFloatComplex *d_primay_beam_J10 = NULL;
     cuFloatComplex *d_primay_beam_J11 = NULL;
     //Same goes for l,m,n coords
-    float *d_ls=NULL;
-    float *d_ms=NULL;
-    float *d_ns=NULL;
+    double *d_ls=NULL;
+    double *d_ms=NULL;
+    double *d_ns=NULL;
 
     if (num_points > 0) {
       printf("\tDoing point components\n");
 
-      float *d_point_ras=NULL;
-      float *d_point_decs=NULL;
[…]
