### 2021-11-01 #3676 @JLBLine
Hi @dfm @mkolopanis @plaplant I reckon I'll need until the end of the week to make the float/double switch work with all the units tests, and then another week to address the rest of the review issues. If you want to pause, I can just notify you all when I've finished everything. I'll open up another issue for the float / double switch so we can tick it off the list at the end.
