### 2021-11-01 JLBLine/WODEN#22 @JLBLine
JOSS review issue 9 (float vs double)
Allow the user to choose between float and double precision. I'll do this by adding a switch that will assign float or double during compilation. This means the same code will always be used, just with different precision. I'll then get CMake to create two executables, `woden_float` and `woden_double`, that `run_woden.py` will be able to call depending on user arguments.

https://github.com/openjournals/joss-reviews/issues/3676
