### 2021-11-03 c389649
Added a switch that uses the compilation flag -DDOUBLE_PRECISION to either compile the code in mostly `float` or mostly `double`. This involved defining my own data types, and so meant I had to touch basically all the code. CMake now creates two executables, `woden_float` and `woden_double`, which `run_woden.py` will call depending on the `--precision` flag. Updated all `ctest` test to be compiled in float or double precision as well. Still need to update the documentation, and maybe have a new `test_installation` test to show float vs double.
