### 2021-11-10 4511514
Made some further changes of select values from float to double to improve accuracy of `woden_float`. Added plenty to the JOSS paper, including the maths of the absolute accuracy estimation. Updated the code that runs said test, and added it to the script that runs all tests in `test_installation` so users can reproduce the plot from the paper.
