+++ joss_paper/paper.bib
@@ -195,3 +195,61 @@ url = {https://academic.oup.com/mnras/article-lookup/doi/10.1093/mnras/stx1547},
 volume = {471},
 year = {2017}
 }
+
+@ARTICLE{HEALPix2005,
+       author = {{G{\'o}rski}, K.~M. and {Hivon}, E. and {Banday}, A.~J. and {Wandelt}, B.~D. and {Hansen}, F.~K. and {Reinecke}, M. and {Bartelmann}, M.},
+        title = "{HEALPix: A Framework for High-Resolution Discretization and Fast Analysis of Data Distributed on the Sphere}",
+      journal = {\apj},
+     keywords = {Cosmology: Cosmic Microwave Background, Cosmology: Observations, Methods: Statistical, Astrophysics},
+         year = 2005,
+        month = apr,
+       volume = {622},
+       number = {2},
+        pages = {759-771},
+          doi = {10.1086/427976},
+archivePrefix = {arXiv},
+       eprint = {astro-ph/0409513},
+ primaryClass = {astro-ph},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2005ApJ...622..759G},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@ARTICLE{FITS2002,
+       author = {{Calabretta}, M.~R. and {Greisen}, E.~W.},
+        title = "{Representations of celestial coordinates in FITS}",
+      journal = {\aap},
+     keywords = {methods: data analysis, techniques: image processing, astronomical data bases: miscellaneous, astrometry, Astrophysics},
+         year = 2002,
+        month = dec,
+       volume = {395},
+        pages = {1077-1122},
+          doi = {10.1051/0004-6361:20021327},
+archivePrefix = {arXiv},
+       eprint = {astro-ph/0207413},
+ primaryClass = {astro-ph},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2002A&A...395.1077C},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@BOOK{TMSthird,
+       author = {{Thompson}, A. Richard and {Moran}, James M. and {Swenson}, George W., Jr.},
+        title = "{Interferometry and Synthesis in Radio Astronomy, 3rd Edition}",
+         year = 2017,
+          doi = {10.1007/978-3-319-44431-4},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2017isra.book.....T},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@INPROCEEDINGS{PAL2013,
+       author = {{Jenness}, T. and {Berry}, D.~S.},
+        title = "{PAL: A Positional Astronomy Library}",
+    booktitle = {Astronomical Data Analysis Software and Systems XXII},
+         year = 2013,
+       editor = {{Friedel}, D.~N.},
+       series = {Astronomical Society of the Pacific Conference Series},
+       volume = {475},
+        month = oct,
+        pages = {307},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2013ASPC..475..307J},
+      adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
