+++ joss_paper/paper.md
@@ -21,28 +21,278 @@ bibliography: paper.bib
 ---
 # Summary
 
-`WODEN` is designed to simulate the response of a class of telescope known as an interferometer, producing output "visibilities" for a given astrophysical sky model. Simulated observations allow us to test other software packages that are designed to calibrate and analyse real interferometric data, including verifying expected behaviour with known inputs, and trialling new sky modelling techniques. The `WODEN` sky model can be specified in dirac-delta like functions on the sky (known in the field as "point sources") elliptical Gaussian models, or built out of "shapelet" basis functions (allowing complicated morphologies to be created). Users are able to input a bespoke layout for the interferometer, vary a number of observational parameters including time of day, length of observation and frequency coverage, and select from a number of predefined primary beams which encode the response of the receiving elements of an interferometer. This allows simulations of a number of telescopes to be undertaken. `WODEN` works with all Stokes $I,Q,U,V$ sky polarisations, simulating telescopes with dual linear polarisations.
+`WODEN` is designed to simulate the response of a class of telescope known as an interferometer, producing output "visibilities" for a given astrophysical sky model. Simulated observations allow us to test other software packages that are designed to calibrate and analyse real interferometric data, including verifying expected behaviour with known inputs, and trialling new sky modelling techniques. The `WODEN` sky model can be specified in dirac-delta like functions on the sky (known in the field as "point sources"), elliptical Gaussian models, or built out of "shapelet" basis functions (allowing complicated morphologies to be created). Users are able to input a bespoke layout for the interferometer, vary a number of observational parameters including time of day, length of observation and frequency coverage, and select from a number of predefined primary beams which encode the response of the receiving elements of an interferometer. This allows simulations of a number of telescopes to be undertaken. `WODEN` works with input Stokes $I,Q,U,V$ polarisations as a sky model, simulating telescopes with dual linear polarisations, outputting linear Stokes polarisations.
 
-The core functionality of `WODEN` is written in CUDA as interferometric simulations are computationally intensive but embarrassingly parallel. The compute performance of CUDA allows for large-scale simulations to be run including emission from all directions in the sky. This is paramount for interferometers with a widefield of view such as the Murchison Widefield Array (MWA, @Tingay2013). A Python wrapper is used to take advantage of community packages such as astropy [@astropy2013, @astropy2018], and to present a user-friendly interface to `WODEN`. Those simulating MWA observations can use the MWA `metafits` file to quickly feed in observational parameters to `WODEN` to match real data.
[…]
