### 2021-11-11 d745659
Updated the JOSS paper and Documentation to better explain how `WODEN` handles polarisaition. A couple of data types that should have been `double` not `user_precision_t` slipped through the net so fixed them up.
