+++ joss_paper/paper.md
@@ -32,10 +32,10 @@ The core functionality of `WODEN` is written in CUDA as interferometric simulati
 An interferometer creates visibilities $V$ by cross-correlating signals detected between pairs of antennas or dishes (baselines), described by coordinates $u,v,w$. Each visibility is sensitive to the entire sky, directions of which we describe by the direction cosines $l,m,n$. Ignoring the antenna response, the full integral can be discretised as
 
 \begin{equation} \label{eq:RIME}
-V_{\mathcal{S}}(u_i,v_i,w_i) = \\ \sum_j \mathcal{S}(l_j,m_j) \exp[-2\pi i(u_il_j + v_im_j + w_i(n_j-1))],
+V_s(u_i,v_i,w_i) = \\ \sum_j \mathcal{S}_s(l_j,m_j) \exp[-2\pi i(u_il_j + v_im_j + w_i(n_j-1))],
 \end{equation}
 
-where $u_i,v_i,w_i$ are the visibility coordinates of the $i^{\mathrm{th}}$ baseline, $l_j$, $m_j$, $n_j$ is the sky position of the $j^{\mathrm{th}}$ component in the sky model, and $\mathcal{S}(l_j,m_j)$ is the flux density of that component in a given Stokes polarisation. `WODEN` simulates dual-linear polarisation antennas, with each
+where $u_i,v_i,w_i$ are the visibility coordinates of the $i^{\mathrm{th}}$ baseline, $l_j$, $m_j$, $n_j$ is the sky position of the $j^{\mathrm{th}}$ component in the sky model, and $\mathcal{S}(l_j,m_j)$ is the flux density of that component in a given Stokes polarisation $s$. `WODEN` simulates dual-linear polarisation antennas, with each
 antenna/station having it's own primary beam shape. I can define the response of a dual polarisation antenna to direction $l,m$ as
 
 $$
@@ -50,10 +50,10 @@ where $g$ are gain terms, $D$ are leakage terms, and $\mathrm{ns}$ refers to nor
 
 \begin{equation}\label{eq:RIME_full}
 \begin{bmatrix}
-V_{12\,{\mathrm{ns}}\,{\mathrm{ns}}}(l,m) \\
-V_{12\,{\mathrm{ns}}\,{\mathrm{ew}}}(l,m) \\
-V_{12\,{\mathrm{ew}}\,{\mathrm{ns}}}(l,m) \\
-V_{12\,{\mathrm{ew}}\,{\mathrm{ew}}}(l,m)
+V_{12\,XX}(l,m) \\
+V_{12\,XY}(l,m) \\
+V_{12\,YX}(l,m) \\
+V_{12\,YY}(l,m)
 \end{bmatrix} =
 \mathbf{J}_1(l,m) \otimes \mathbf{J}_2^*(l,m)
 \begin{bmatrix}
@@ -70,14 +70,14 @@ V_{12\,V}(l,m)
 \end{bmatrix}
 \end{equation}
 
+
+
 where $*$ denotes a complex conjugate, and $\otimes$ an outer product (the result
-of this outer product is written explicitly in the `WODEN` documentation [here](https://woden.readthedocs.io/en/joss_review/operating_principles/visibility_calcs.html)). For each baseline, frequency, and time step, `WODEN` calculates all four linear polarisations[^1] as defined above for all directions $l,m$ in the sky model, and then sums over sky direction, to produce four full-sky linear Stokes polarisation visibilities per baseline/frequency/time.
[…]
