+++ test_installation/absolute_accuracy/create_array_layouts.py
@@ -1,4 +1,14 @@
 import numpy as np
+import palpy as pal
+from astropy.time import Time, TimeDelta
+from astropy.coordinates import EarthLocation
+from astropy import units as u
+
+DD2R = np.pi/180.0
+
+VELC  = 299792458.0
+MWA_LAT = -26.703319
+MWA_LONG = 116.67081524
 
 def calc_b(phi_simple, num_mult):
     """Given the target angle `phi_simple` (radians), and
@@ -15,19 +25,159 @@ def calc_b(phi_simple, num_mult):
 
     return b
 
-##Known list of angles that have predictable sin/cos outputs
-phi_simples = [0.0, np.pi/6, np.pi/4, np.pi/3, np.pi/2, 2*np.pi/3,
-               3*np.pi/4, 5*np.pi/6, np.pi, 7*np.pi/6, 5*np.pi/4]
+def enh2xyz(east, north, height, latitiude):
+    """Convert local e,n,h to local x,y,z coords"""
+    sl = np.sin(latitiude)
+    cl = np.cos(latitiude)
+    X = -north*sl + height*cl
+    Y = east
+    Z = north*cl + height*sl
 
-##For known angles, calculate l,m,n coords and check they are legit
-for ind, phi_simple in enumerate(phi_simples):
+    return X, Y, Z
 
-    for num_mult in [1, 10, 100, 1000, 10000]:
+def get_uvw(xdiff, ydiff, zdiff, dec0, ha0):
+    """Get the u,v,w given a baseline length in local X, Y, Z coords,
+    and a phase centre ha0, dec0"""
+    u = np.sin(ha0)*xdiff + np.cos(ha0)*ydiff
+    v = -np.sin(dec0)*np.cos(ha0)*xdiff + np.sin(dec0)*np.sin(ha0)*ydiff + np.cos(dec0)*zdiff
+    w = np.cos(dec0)*np.cos(ha0)*xdiff - np.cos(dec0)*np.sin(ha0)*ydiff + np.sin(dec0)*zdiff
 
-        b = calc_b(phi_simple, num_mult)
+    return u,v,w
 
-        with open(f"array_layouts/array_layout_ang{ind:02d}_n{num_mult:05d}.txt", 'w') as outfile:
-            outfile.write(f"{b} {b} {b}\n")
-            outfile.write("0.0 0.0 0.0")
+def rotation_applied_by_woden(current_lst, mjd, latitude, ra0 = 0.0, dec0 = 0.0):
 
-        # phi_outcome = 2*np.pi*b*(l + m + n - 1)
+    current_time_rotation = np.zeros((3,3))
+    current_time_rotation[0,0] = np.cos(current_lst)
+    current_time_rotation[0,1] = -np.sin(current_lst)
+    current_time_rotation[1,0] = np.sin(current_lst)
+    current_time_rotation[1,1] = np.cos(current_lst)
+    current_time_rotation[2,2] = 1
+
+
+    # print("INPUTS INTO PALPRENUT", 2000.0, mjd)
+
+    pal_rotation_matrix = pal.prenut(2000.0, mjd)
+    pal_rotation_matrix = np.transpose(pal_rotation_matrix)
+
+    # print(pal_rotation_matrix)
+
+    v1 = pal.dcs2c(current_lst, latitude)
+    v2 = pal.dmxv(pal_rotation_matrix, v1)
+    J2000_lst, J2000_lat = pal.dcc2s(v2)
+
+    J2000_lst = pal.dranrm(J2000_lst)
+
+
+    print("Current_LST, Current_Latitude", current_lst, latitude)
+
+    if J2000_lst > np.pi:
[…]
