### 2021-11-11 f07a914
Putting palpy into requirements.txt killed ReadTheDocs, so setup a separate file to send to ReadTheDocs
