### 2021-11-11 2017c55
ReadTheDocs refuses to not ues requirements.txt I think, so try a new set of text files
