+++ docs/sphinx/requirements.txt
@@ -3,5 +3,3 @@ breathe
 astropy
 numpy
 pyerfa
-palpy
-matplotlib
