### 2021-11-11 JLBLine/WODEN#21 @JLBLine
As of commit 2017c55 I have made the following edits to the paper and the Documentation:

 - *(known in the field as "point sources) -> this parenthetical should be followed by a comma.*
     - Fixed

 - *Astropy Collaboration citation -> should this be a parenthetical citation?*
    - Fixed, was a typo

 - *Under a formalism like the above -> I understand what is meant by this sentence but it is difficult to parse. Perhaps something along the lines of "Under this discrete sky formalism, upwards of j>=25 x 10^6 or more components…"*
    - I've made the suggested change

 - *The mathematical expression of the radio interferometer measurement equation (RIME) in the docs and in the paper should include reference to polarization.*
    - I've added plenty of text and equations defining how polarisation is handled in both the paper, and in the Documentation [here](https://woden.readthedocs.io/en/joss_review/operating_principles/visibility_calcs.html)

 - *It might be worth mentioning exactly how the pixelization of the sphere is done (e.g., a regular grid in (l, m), HEALPix pixels, something else?), as it affects how things like the sky model and beam interpolation are handled.*
     - I've added text discussing how the pixelisation on the sphere is chosen by the user via them creating their own sky model

 - *Although there are benchmarking results included in both the reference paper (Line et al. 2020) and some of the documentation pages, it might be nice to include it in the "Example application" section as well, to highlight the speed of the calculation.*
     - I've added benchmarking simulation outcomes for two different GPU cards into the paper
