### 2021-11-11 JLBLine/WODEN#22 @JLBLine
As of commit d745659cf36efe9ee6750fd075b5597b7ad0de00, the "double" or "float" choice is working in the code and unit testing, and has been written up into the JOSS paper. Still need to re-run all benchmarks, and need to update the Documentation in places.
