### 2021-11-11 #3676 @JLBLine
Hey @dfm @mkolopanis @plaplant, the float vs double switch is working now, and I've updated the unit/integration tests accordingly. I've update the JOSS paper [here](https://github.com/JLBLine/WODEN/blob/joss_review/joss_paper/paper.pdf) with the new estimate on accuracy, and better explanations on how I handle polarisation.

Still have a few of the repo issues to catch up on, so the work will spill into next week, but I think the code itself is in it's final form for this paper. Mostly Documentation that needs to catch up.
