### 2021-11-18 aae1888
Massive effort to document all the unit tests. Worked out how to actually test 64 bit precision using Unity, so update all of the C and CUDA units tests to use TEST_ASSERT_DOUBLE where appropriate. Through this new testing, realised a float isn't accurate enough to store numbers in the 100e+6 Hz to 1Hz precision, so swapped all frequencies to 64 bit precision for both the FLOAT and DOUBLE versions of the code
