+++ cmake_testing/chunk_sky_model/fill_chunks_common.c
@@ -94,7 +94,7 @@ catsource_t * make_sky_model(int num_points, int num_gauss,
   //Populate POINT intrinsic properties
   cropped_src->point_ras = index_point_array_double;
   cropped_src->point_decs = index_point_array_double;
-  cropped_src->point_ref_freqs = index_point_array;
+  cropped_src->point_ref_freqs = index_point_array_double;
   cropped_src->point_ref_stokesI = index_point_array;
   cropped_src->point_ref_stokesQ = index_point_array;
   cropped_src->point_ref_stokesU = index_point_array;
@@ -131,7 +131,7 @@ catsource_t * make_sky_model(int num_points, int num_gauss,
 
   cropped_src->gauss_ras = index_gauss_array_double;
   cropped_src->gauss_decs = index_gauss_array_double;
-  cropped_src->gauss_ref_freqs = index_gauss_array;
+  cropped_src->gauss_ref_freqs = index_gauss_array_double;
   cropped_src->gauss_ref_stokesI = index_gauss_array;
   cropped_src->gauss_ref_stokesQ = index_gauss_array;
   cropped_src->gauss_ref_stokesU = index_gauss_array;
@@ -161,7 +161,7 @@ catsource_t * make_sky_model(int num_points, int num_gauss,
 
   cropped_src->shape_ras = index_shape_array_double;
   cropped_src->shape_decs = index_shape_array_double;
-  cropped_src->shape_ref_freqs = index_shape_array;
+  cropped_src->shape_ref_freqs = index_shape_array_double;
   cropped_src->shape_ref_stokesI = index_shape_array;
   cropped_src->shape_ref_stokesQ = index_shape_array;
   cropped_src->shape_ref_stokesU = index_shape_array;
@@ -195,13 +195,13 @@ void free_sky_model(catsource_t *cropped_src) {
   //These point to the unique arrays we made in `make_sky_model`
   //so freeing just these is enough
   free(cropped_src->point_ras);
-  free(cropped_src->point_ref_freqs);
+  free(cropped_src->point_ref_stokesI);
   free(cropped_src->point_azs);
   free(cropped_src->gauss_ras);
-  free(cropped_src->gauss_ref_freqs);
+  free(cropped_src->gauss_ref_stokesI);
   free(cropped_src->gauss_azs);
   free(cropped_src->shape_ras);
-  free(cropped_src->shape_ref_freqs);
+  free(cropped_src->shape_ref_stokesI);
   free(cropped_src->shape_azs);
   free(cropped_src->shape_coeffs);
 
@@ -266,6 +266,9 @@ void check_pointgauss_chunking(int chunk_ind, int comps_per_chunk,
 
     TEST_ASSERT_EQUAL_INT(expected_n_points, temp_cropped_src->n_points);
 
+    //Some things are a choice between float and double at compilation, and
+    //some things are always set to double, so setup two different expectation
+    //arrays
     user_precision_t *expec_index_point_array = malloc(expected_n_points*sizeof(user_precision_t));
     make_index_array(expec_index_point_array, expected_n_points, * point_accum);
 
@@ -273,16 +276,20 @@ void check_pointgauss_chunking(int chunk_ind, int comps_per_chunk,
     make_repeat_array(expec_repeat_point_array, expected_n_points,
                       num_time_steps, * point_accum);
 
+    double *expec_index_point_array_double = malloc(expected_n_points*sizeof(double));
[…]
