### 2021-11-18 JLBLine/WODEN#17 @JLBLine
As of commit aae18882033bb570be006fbdfa2570d929396353 I have a Section called [What do the tests actually do?](https://woden.readthedocs.io/en/joss_review/testing/cmake_testing.html#what-do-the-tests-actually-do), which contains subsection links that detail all the unit tests made for each `src` file.

https://github.com/openjournals/joss-reviews/issues/3676 @mkolopanis @plaplant I think this is enough to address this issue
