### 2021-11-18 39f2dea
Added a contribution guide
