### 2021-11-18 JLBLine/WODEN#20 @JLBLine
With commit 39f2deab0268adfbeda28fb7166dbb1bab2d79b7 I've added the [CONTRIBUTION_GUIDE.md](https://github.com/JLBLine/WODEN/blob/joss_review/CONTRIBUTION_GUIDE.md) and linked to that in the `README.md`.
@mkolopanis @plaplant I think this addresses the issue
