### 2021-11-18 a51b7ea
Small GPU can't handle DOUBLE for a decent number of SHAPELET components, so added --precision=float for a couple of the installation tests
