### 2021-11-19 8b13952
Updated some benchmarks and tidied up the JOSS paper a little.
