+++ joss_paper/paper.bib
@@ -253,3 +253,41 @@ archivePrefix = {arXiv},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2013ASPC..475..307J},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
+
+
+Marten van Kerkwijk; Erik Tollerud; Julien Woillez; Thomas Robitaille; E. M. Bray; Antonio Valentino; Brigitta Sipőcz; Michael Droettboom; Christoph Deil; Michael Seifert; Simon Conseil; Tom Aldcroft; Adrian Price-Whelan; StuartLittlefair; P. L. Lim; Bernardo Sulzbach; Chris Beaumont; Daria Cara; Devin Crichton; Matt Davis; Miguel de Val-Borro; Sergio Pascual; Tomas Babej; Vatsala Swaroop; Vishnunarayan K I; Jani Šumak
+
+@misc{pyerfa,
+  author    = {{van Kerkwijk}, Marten and {Tollerud}, Erik and {Woillez}, Julien and {Robitaille}, Thomas and {Bray}, E. M. and {Valentino}, Antonio and {Sipocz}, Brigitta and  {Droettboom}, Michael and {Deil}, Christoph and {Seifert}, Michael and {Conseil}, Simon and {Aldcroft}, Tom and {Price-Whelan}, Adrian and {StuartLittlefair}, and {Lim}, P. L. and  {Sulzbach}, Bernardo and {Beaumont}, Chris and {Cara}, Daria and {Crichton}, Devin and {Davis}, Matt and {de Val-Borro}, Miguel, and {Pascual}, Sergio and {Babej}, Tomas and {Swaroop}, Vatsala and {Vishnunarayan}, K. I. and {Sumak}, Jani},
+  title     = {pyerfa},
+  version   = {2.0.0},
+  publisher = {Zenodo},
+  month     = jul,
+  year      = 2020,
+  doi       = {10.5281/zenodo.3940699},
+  url       = {https://doi.org/10.5281/zenodo.3940699}
+}
+
+
+@article{Lanman2019,
+  doi = {10.21105/joss.01234},
+  url = {https://doi.org/10.21105/joss.01234},
+  year = {2019},
+  publisher = {The Open Journal},
+  volume = {4},
+  number = {37},
+  pages = {1234},
+  author = {Adam E. Lanman and Bryna J. Hazelton and Daniel C. Jacobs and Matthew J. Kolopanis and Jonathan C. Pober and James E. Aguirre and Nithyanandan Thyagarajan},
+  title = {pyuvsim: A comprehensive simulation package for radio interferometers in python},
+  journal = {Journal of Open Source Software}
+}
+
+@INPROCEEDINGS{OSKAR,
+  author={Mort, Benjamin J. and Dulwich, Fred and Salvini, Stefano and Adami, Kristian Zarb and Jones, Mike E.},
+  booktitle={2010 IEEE International Symposium on Phased Array Systems and Technology},
+  title={OSKAR: Simulating digital beamforming for the SKA aperture array},
+  year={2010},
+  volume={},
+  number={},
+  pages={690-694},
+  doi={10.1109/ARRAY.2010.5613289}}
