+++ joss_paper/paper.md
@@ -23,13 +23,13 @@ bibliography: paper.bib
 
 `WODEN` is designed to simulate the response of a class of telescope known as an interferometer, producing output "visibilities" for a given astrophysical sky model. Simulated observations allow us to test other software packages that are designed to calibrate and analyse real interferometric data, including verifying expected behaviour with known inputs, and trialling new sky modelling techniques. The `WODEN` sky model can be specified in dirac-delta like functions on the sky (known in the field as "point sources"), elliptical Gaussian models, or built out of "shapelet" basis functions (allowing complicated morphologies to be created). Users are able to input a bespoke layout for the interferometer, vary a number of observational parameters including time of day, length of observation and frequency coverage, and select from a number of predefined primary beams which encode the response of the receiving elements of an interferometer. This allows simulations of a number of telescopes to be undertaken. `WODEN` works with input Stokes $I,Q,U,V$ polarisations as a sky model, simulating telescopes with dual linear polarisations, outputting linear Stokes polarisations.
 
-The core functionality of `WODEN` is written in CUDA as interferometric simulations are computationally intensive but embarrassingly parallel. The compute performance of CUDA allows for large-scale simulations to be run including emission from all directions in the sky. This is paramount for interferometers with a widefield of view such as the Murchison Widefield Array (MWA, @Tingay2013). A Python wrapper is used to take advantage of community packages such as astropy [@astropy2013; @astropy2018], and to present a user-friendly interface to `WODEN`. Those simulating MWA observations can use the MWA `metafits` file to quickly feed in observational parameters to `WODEN` to match real data.
+The core functionality of `WODEN` is written in CUDA as interferometric simulations are computationally intensive but embarrassingly parallel. The compute performance of CUDA allows for large-scale simulations to be run including emission from all directions in the sky. This is paramount for interferometers with a widefield of view such as the Murchison Widefield Array (MWA, @Tingay2013). A Python wrapper is used to take advantage of community packages such as [astropy](https://www.astropy.org/) [@astropy2013; @astropy2018] and [pyerfa](https://pypi.org/project/pyerfa/) [@pyerfa] and to present a user-friendly interface to `WODEN`. Those simulating MWA observations can use the MWA `metafits` file to quickly feed in observational parameters to `WODEN` to match real data.
 
-`WODEN` can be run to two levels of precision: a "float" precision (which uses a mix of 32 and 64 bit floating precision), and a "double" (which uses nearly entirely 64 bit precision). In Section [Estimation of accuracy and computational speed](#Estimation of accuracy and computational speed) `WODEN` is shown to produce visibilities to within 0.2% of expected values when running in "float" mode, and 0.000002% in "double" mode, for baselines of length <= 10km.
+`WODEN` can be run to two levels of precision: a `woden_float` precision (which uses a mix of 32 and 64 bit floating precision), and a `woden_double` (which uses nearly entirely 64 bit precision). In Section [Estimation of accuracy and computational speed](#Estimation of accuracy and computational speed) `WODEN` is shown to produce visibilities to within 0.2% of expected values when running in `woden_float` mode, and 0.000002% in `woden_double` mode, for baselines of length <= 10km.
 
 # Underlying methodolgy
 
[…]
