+++ examples/MWA_EoR1/MWA_EoR1_simulation.sh
@@ -9,12 +9,10 @@ do
     --freq_res=80e+3 --time_res=8.0 \
     --cat_filename=woden-srclist_pumav3.txt \
     --metafits_filename=../metafits/1136380296_metafits_ppds.fits \
-    --band_nums=1 \
+    --band_nums=1,2,3,4,5 \
     --output_uvfits_prepend=./data/MWA_EoR1_${precision} \
     --primary_beam=MWA_FEE \
     --sky_crop_components \
     --precision=${precision}
-    
-done
 
-#,2,3,4,5
+done
