+++ docs/sphinx/examples/example_simulations.rst
@@ -17,8 +17,10 @@ The examples are:
    MWA_EoR1_sim
    eda2_haslam_sim
 
-**Fornax A simulation** - two examples that compare a point/Gaussian model to a shapelet model. This serves as a general introduction on how to simulate an MWA observation using a ``metafits`` file, with a few extra commands
+**Fornax A simulation** - two examples that compare a point/Gaussian model to a shapelet model. This serves as a general introduction on how to simulate an MWA observation using a ``metafits`` file, with a few extra commands. It also serves as a comparison of running with ``woden_float`` and ``woden_double``.
 
 **MWA EoR1 simulation** - demonstrates using a larger (>300,000) source catalogue
 
 **EDA2 Haslam Map simulation** - this demonstrates using ``WODEN`` without a ``metafits`` file, using a text file to describe the array layout, and using the ``EDA2`` beam.
+
+.. warning:: If you have a GPU with small amounts of RAM (say 2GB) some of these simulations won't work to DOUBLE precision, you won't have enough memory. You can add the ``--precision=float`` argument to switch to a lower memory requirement (for the loss of accuracy).
