### 2021-11-19 JLBLine/WODEN#22 @JLBLine
@mkolopanis @plaplant as of commit 8b139527cbfbc934a2f92442a3b3778710c3ffff the documentation should be caught up to the float/double precision comparisons so I think this issue has been resolved now
