### 2021-11-19 5996d51
Added bench mark results from the EDA2 simulation for the new `woden_float` and `woden_double` versions
