+++ docs/sphinx/examples/eda2_haslam_sim.rst
@@ -15,7 +15,7 @@ To run the command, do::
 
   $ ./EDA2_haslam_simulation.sh
 
-which took (XX hours, benchmarking is running) on my machine (this is running in DOUBLE precision). The command run is::
+which took 2 hours 11 mins on my machine (this is running in DOUBLE precision, it takes 1 hour 15 mins at FLOAT precision). This simulates 393,216 point sources for an array of 255 antennas. The command run is::
 
   run_woden.py \
     --ra0=74.79589467 --dec0=-27.0 \
