### 2021-11-24 e00e09f
Was making a small error in writing out the times in the DATE array in the output uvfits, need to add half a time step
