### 2021-11-26 a0229bf
Needed to update the unit test for the DATE array in the uvfits file after adding half a time step resolution the old values were incorrect. All CTests run now.
