+++ cmake_testing/FEE_primary_beam/test_multifreq_RTS_MWAFEEInit.h
@@ -1,0 +1,82 @@
+
+//First set of delays are the zenith delays
+double expec_freq1delay1_zen[] = {-0.00225198, -0.00233495, -0.00623328,
+  -0.00646973, -0.00000003, 0.00000000, 0.00000010, 0.00000001, -0.00225411,
+  -0.00233764, -0.00624081, -0.00647886, -0.00000003, 0.00000000, 0.00000010,
+  0.00000001, -0.00225625, -0.00234033, -0.00624836, -0.00648800, -0.00000003,
+  0.00000000, 0.00000010, 0.00000001, -0.00225838, -0.00234301, -0.00625591,
+  -0.00649717, -0.00000003, 0.00000000, 0.00000010, 0.00000001, -0.00226051,
+  -0.00234569, -0.00626348, -0.00650635, -0.00000003, 0.00000000, 0.00000011,
+  0.00000001, -0.00226264, -0.00234837, -0.00627107, -0.00651554, -0.00000003,
+  0.00000000, 0.00000011, 0.00000001, -0.00226476, -0.00235105, -0.00627867,
+  -0.00652476, -0.00000003, 0.00000000, 0.00000011, 0.00000001, -0.00226688,
+  -0.00235372, -0.00628628, -0.00653398, -0.00000002, 0.00000000, 0.00000011,
+  0.00000001, -0.00226899, -0.00235639, -0.00629391, -0.00654323, -0.00000002,
+  0.00000000, 0.00000011, 0.00000001, -0.00227110, -0.00235906, -0.00630155,
+  -0.00655249, -0.00000002, 0.00000000, 0.00000011, 0.00000001, -0.00227321,
+  -0.00236173, -0.00630920, -0.00656176, -0.00000002, 0.00000000, 0.00000011,
+  0.00000001, -0.00227532, -0.00236439, -0.00631687, -0.00657105, -0.00000001,
+  0.00000000, 0.00000011, 0.00000000, -0.00227742, -0.00236704, -0.00632455,
+  -0.00658036, -0.00000001, 0.00000000, 0.00000011, 0.00000000, -0.00227952,
+  -0.00236969, -0.00633225, -0.00658968, -0.00000001, 0.00000000, 0.00000011,
+  0.00000000, -0.00228161, -0.00237234, -0.00633996, -0.00659902, -0.00000001,
+  0.00000000, 0.00000011, 0.00000000, -0.00228370, -0.00237499, -0.00634769,
+  -0.00660837, -0.00000000, 0.00000000, 0.00000011, 0.00000000, -0.00228579,
+  -0.00237763, -0.00635543, -0.00661774, -0.00000000, 0.00000000, 0.00000010,
+  0.00000000, -0.00228787, -0.00238026, -0.00636318, -0.00662713, 0.00000000,
+  0.00000000, 0.00000010, -0.00000000, -0.00228995, -0.00238289, -0.00637095,
+  -0.00663653, 0.00000000, 0.00000001, 0.00000010, -0.00000000, -0.00229203,
+  -0.00238551, -0.00637873, -0.00664594, 0.00000001, 0.00000001, 0.00000010,
+  -0.00000000, -0.00229410, -0.00238813, -0.00638653, -0.00665537, 0.00000001,
+  0.00000001, 0.00000010, -0.00000000, -0.00229617, -0.00239075, -0.00639434,
+  -0.00666482, 0.00000001, 0.00000001, 0.00000009, -0.00000000, -0.00229823,
+  -0.00239336, -0.00640217, -0.00667428, 0.00000002, 0.00000001, 0.00000009,
+  -0.00000001, -0.00230029, -0.00239596, -0.00641001, -0.00668376, 0.00000002,
[…]
