### 2021-12-02 JLBLine/WODEN#20 @mkolopanis
Thanks for the contribution guide. This addresses this issue as far as I am concerned.
