### 2021-12-14 84f4a55
edits for JOSS review
