+++ joss_paper/paper.bib
@@ -101,7 +101,7 @@ year = {2021}
          {Weaver}, B.~A. and {Whitmore}, J.~B. and {Woillez}, J. and
          {Zabalza}, V. and {Astropy Contributors}},
         title = "{The Astropy Project: Building an Open-science Project and Status of the v2.0 Core Package}",
-      journal = {\aj},
+      journal = {The Astronomical Journal},
      keywords = {methods: data analysis, methods: miscellaneous, methods: statistical, reference systems, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2018,
         month = sep,
@@ -135,7 +135,7 @@ archivePrefix = {arXiv},
          {Pierfederici}, Francesco and {Pontzen}, Andrew and {Ptak}, Andy and
          {Refsdal}, Brian and {Servillat}, Mathieu and {Streicher}, Ole},
         title = "{Astropy: A community Python package for astronomy}",
-      journal = {\aap},
+      journal = {Astronomy \& Astrophysics},
      keywords = {methods: data analysis, methods: miscellaneous, virtual observatory tools, Astrophysics - Instrumentation and Methods for Astrophysics},
          year = 2013,
         month = oct,
@@ -199,7 +199,7 @@ year = {2017}
 @ARTICLE{HEALPix2005,
        author = {{G{\'o}rski}, K.~M. and {Hivon}, E. and {Banday}, A.~J. and {Wandelt}, B.~D. and {Hansen}, F.~K. and {Reinecke}, M. and {Bartelmann}, M.},
         title = "{HEALPix: A Framework for High-Resolution Discretization and Fast Analysis of Data Distributed on the Sphere}",
-      journal = {\apj},
+      journal = {The Astrophysical Journal},
      keywords = {Cosmology: Cosmic Microwave Background, Cosmology: Observations, Methods: Statistical, Astrophysics},
          year = 2005,
         month = apr,
@@ -217,7 +217,7 @@ archivePrefix = {arXiv},
 @ARTICLE{FITS2002,
        author = {{Calabretta}, M.~R. and {Greisen}, E.~W.},
         title = "{Representations of celestial coordinates in FITS}",
-      journal = {\aap},
+      journal = {Astronomy \& Astrophysics},
      keywords = {methods: data analysis, techniques: image processing, astronomical data bases: miscellaneous, astrometry, Astrophysics},
          year = 2002,
         month = dec,
