### 2022-01-04 2813344
Updated the JOSS paper with a link to Version 1.1, which is the release containing the changes made during review. I re-ran the code to make the accuracy plot with this release version and updated the figure in the paper. The conclusions from the plot remain the same.
