### 2022-01-04 #3676 @dfm
@JLBLine: welcome back! Thanks for your updates on the manuscript. Here are the last steps that I'll need from you:

1. Take one last read through the manuscript to make sure that you're happy with it (it's harder to make changes later!), especially your name and affiliation.
2. Confirm that the version number including all the changes as part of this review is `v1.1.0`.

Thanks!
