### 2022-01-05 d907e5f
Updated my affiliation on the JOSS paper
