### 2022-01-05 #3676 @JLBLine
Ok @dfm I believe I've made a Zenodo release correctly, it's been given the DOI `10.5281/zenodo.5819254`. Link to the Zenodo if you need it: https://zenodo.org/record/5819254#.YdUPWntBzJk
