### 2022-01-05 #3676 @kthyng
@JLBLine -

* Version and Zenodo look good ✅
* paper:
- "python" isn't capitalized in Lanman et al 2019; please look through all references to see if other corrections should also be made.
