### 2022-01-06 39d0082
Got rid of capital on python for the JOSS paper
