### 2022-01-06 #3676 @dfm
@JLBLine: That's not what @kthyng meant! She was suggesting that the title of that paper should have a capital `P` (I note that the published version does, in fact, have a lowercase `p`: https://joss.theoj.org/papers/10.21105/joss.01234) but you can update the bib file to:

```diff
-  title = {pyuvsim: A comprehensive simulation package for radio interferometers in python},
+  title = {{pyuvsim}: A comprehensive simulation package for radio interferometers in {P}ython},
```

I also think you should revert your previous change: Python should, in general, be capitalized.
