### 2022-01-06 272cad2
Put the capital P back on Python for the JOSS paper
