+++ joss_paper/paper.bib
@@ -278,7 +278,7 @@ Marten van Kerkwijk; Erik Tollerud; Julien Woillez; Thomas Robitaille; E. M. Bra
   number = {37},
   pages = {1234},
   author = {Adam E. Lanman and Bryna J. Hazelton and Daniel C. Jacobs and Matthew J. Kolopanis and Jonathan C. Pober and James E. Aguirre and Nithyanandan Thyagarajan},
-  title = {pyuvsim: A comprehensive simulation package for radio interferometers in python},
+  title = {{pyuvsim}: A comprehensive simulation package for radio interferometers in {P}ython},
   journal = {Journal of Open Source Software}
 }
 
