### 2022-01-06 #3676 @JLBLine
Thanks @dfm, my mistake. I've made the bib change for Lanman et al 2019 as suggested, and reverted python back to Python in the text. To the best of my knowledge all is now as it should be @kthyng.
